src/ingress/pyproject.toml
[project]
name = "ingress"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.133.1",
    "jinja2>=3.1.6",
    "json-stream>=2.4.1",
    "psycopg2-binary>=2.9.11",
    "python-multipart>=0.0.22",
]
